django/requirements.txt
asgiref==3.3.4
attrs==20.3.0
Automat==20.2.0
autopep8==1.5.6
cffi==1.14.5
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
parsel==1.6.0
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyDispatcher==2.0.5
pyOpenSSL==20.0.1
python-decouple==3.4
pytz==2021.1
queuelib==1.6.1
Scrapy==2.5.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==21.2.0
w3lib==1.22.0
zope.interface==5.4.0
